Minor cleanup: Use size_t for struct size cast (CVS 4940)
FossilOrigin-Name: 618df68b8b78d376d30cea79a273fd39140f5033

+++ src/vdbemem.c
@@ -508,7 +508,7 @@ int sqlite3VdbeMemTooBig(Mem *p){
 /*
 ** Size of struct Mem not including the Mem.zMalloc member.
 */
-#define MEMCELLSIZE (int)(&(((Mem *)0)->zMalloc))
+#define MEMCELLSIZE (size_t)(&(((Mem *)0)->zMalloc))
 
 /*
 ** Make an shallow copy of pFrom into pTo.  Prior contents of
